seperate entrypoints for on prem and on cloud
